feat: create BedrockRendererWebpackBase
Most of the BedrockReactRenderer logic went in here.
