+++ bedrock/src/index.js
@@ -1,9 +1,13 @@
 const { bedrockEvents, EVENTS } = require('./server/events');
 const log = require('./cli/log');
-const BedrockRendererBase = require('./server/renderer-base');
+const { BedrockRendererBase } = require('./server/renderer-base');
+const {
+  BedrockRendererWebpackBase,
+} = require('./server/renderer-webpack-base');
 
 module.exports = {
   BedrockRendererBase,
+  BedrockRendererWebpackBase,
   bedrockEvents,
   EVENTS,
   log,
